// web/shared/api/types.gen.ts
export type PublicUserSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * Username
     */
    username: string;
    /**
     * Avatar
     */
    avatar?: string;
    /**
     * Banner
     */
    banner?: string;
};
// ...
export type UserSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * Username
     */
    username: string;
    /**
     * Email
     */
    email: string;
    /**
     * Avatar
     */
    avatar?: string;
    /**
     * Banner
     */
    banner?: string;
};